ovdserver/backends/APE/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.6.0
-e git+https://github.com/shenyunhang/APE.git@8c4920e2014d4818ec3ecafb7b526d896a659d1c#egg=ape&subdirectory=../../../../../../../../../external/OVDServer/ovdserver/backends/APE/shenyunhang/APE
async-timeout==4.0.3
attrs==24.2.0
autoflake==1.7.8
black==22.3.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.11
datasets==3.0.1
detectron2 @ git+https://github.com/facebookresearch/detectron2@017abbfa5f2c2a2afa045200c2af9ccf2fc6227f
detrex @ git+https://github.com/IDEA-Research/detrex@776058ec229be37a5ff2a2b0bb54133bdd5da663
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.2
fairscale==0.4.13
fastapi==0.115.0
filelock==3.16.1
flake8==3.8.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.1
h11==0.14.0
huggingface-hub==0.25.1
hydra-core==1.1.2
idna==3.10
iniconfig==2.0.0
iopath==0.1.9
isort==4.3.21
kiwisolver==1.4.7
lvis==0.5.3
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.8.4
mccabe==0.6.1
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
numpy==1.22.4
omegaconf==2.1.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pybind11==2.13.6
pycocotools==2.0.8
pycodestyle==2.6.0
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==2.2.0
pyparsing==3.1.4
pyre-extensions==0.0.23
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.7.3
sentry-sdk==2.14.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.38.6
submitit==1.5.2
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
timm==1.0.9
tokenizers==0.13.3
tomli==2.0.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.66.5
transformers==4.32.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
wandb==0.18.1
wcwidth==0.2.13
Werkzeug==3.0.4
xformers @ file:///home/llq/workspace/code/APE/xformers/xformers
xxhash==3.5.0
yacs==0.1.8
yarl==1.15.5
